fix(pg-delta): enforce public shadow identity
Move observed PostgreSQL identity into a neutral module and require planSchemaFiles to reject the target database or a same-lineage isolated shadow before profile resolution and SQL loading. Preserve the CLI safety exports and early preflight.

RED: same-database and isolated same-lineage frontend calls resolved instead of rejecting; identity permission failure reached profile resolution. GREEN: focused frontend/CLI suites pass 16/16, identity/schema unit suites pass 22/22, and check-types passes.
