+++ .changeset/safety-batch-fixes.md
@@ -10,4 +10,4 @@ Seven safety fixes across the plan, proof, and SQL-file frontends:
 - **Scratch-mode shadow loads contain cluster-object leaks.** In `databaseScratch` mode, `loadSqlFiles` now preflights every file for cluster-global DDL (roles/memberships) before executing anything, and runs a best-effort restore (drop created roles, invert membership deltas) on all exit paths — including error paths that previously skipped the leak check — so a committed `CREATE ROLE` or a DO-block dynamic role no longer survives a failed load.
 - **The proof loop now covers renamed tables.** Accepted table renames are stamped on the plan artifact; the proof maps the old table to its new name so its row count and (when the schema signature is unchanged) content fingerprint are verified as CHECKED, instead of being skipped as recreated and left with zero data-preservation coverage.
 - **Rendered migrations enforce destruction metadata integrity.** The renderer now rejects actions that destroy a table, materialized view, column, or composite attribute while claiming `dataLoss:none`, even when destructive rendering was explicitly allowed. Accepted same-action renames remain exempt, matching apply and prove.
-- **Proof clones and declarative shadows now use observed PostgreSQL identity.** CLI plans stamp domain-separated opaque hashes of `pg_control_system().system_identifier` (the physical-replication lineage) and the source database OID. `prove` rejects source aliases and same-lineage cluster-scoped clones before extraction or DDL; legacy/direct-library artifacts and servers that deny identity observation require the explicit `--allow-unverified-source-identity` seatbelt, which cannot override a confirmed match. Permission denials recommend the required grant and re-plan, while servers where `pg_control_system()` is unavailable report that unsupported capability without impossible grant advice. Explicit schema shadows use the same transport-independent checks with connection-specific permission diagnostics, while an isolation request now requires an explicit shadow from a different lineage. Physical/base-backup clones retain the source identity and are intentionally unsupported.
+- **Proof clones and declarative shadows now use observed PostgreSQL identity.** CLI plans stamp domain-separated opaque hashes of `pg_control_system().system_identifier` (the physical-replication lineage) and the source database OID. `prove` rejects source aliases and same-lineage cluster-scoped clones before extraction or DDL; legacy/direct-library artifacts and servers that deny identity observation require the explicit `--allow-unverified-source-identity` seatbelt, which cannot override a confirmed match. Permission denials recommend the required grant and re-plan, while servers where `pg_control_system()` is unavailable report that unsupported capability without impossible grant advice. Explicit schema shadows use the same transport-independent checks with connection-specific permission diagnostics, while an isolation request now requires an explicit shadow from a different lineage. The public `planSchemaFiles` frontend enforces the same database and lineage invariants before profile resolution or shadow loading, rather than trusting its `isolatedShadow` assertion and disabling shared-cluster containment on an unsafe pool. Physical/base-backup clones retain the source identity and are intentionally unsupported.
